requirements.txt
streamlit
pandas
numpy
torch
re
transformers 
sentence_transformers
sklearn.metrics.pairwise